requirements.txt
langchain-core>=0.1.34
langchain-community==0.0.29
langchain-openai==0.0.2.post1
openai==1.6.1
python-dotenv==0.19.2
requests==2.28.1
yfinance==0.1.87
pydantic==2.5.2
aiohttp==3.9.1